Refactor eval runner into harbor
