+++ cmd/eval_test.go
@@ -3,24 +3,20 @@ package cmd
 import (
 	"bytes"
 	"errors"
-	"fmt"
 	"io"
-	"os"
-	"path/filepath"
 	"reflect"
 	"strings"
 	"testing"
 
-	"github.com/go-git/go-git/v5"
-
-	"github.com/runmedev/runme/v3/internal/ansi"
+	"github.com/runmedev/runme/v3/internal/harbor"
 )
 
-type recordedCommand struct {
-	name       string
-	args       []string
-	workingDir string
-	env        []string
+type fakeEvalRunner struct {
+	err error
+}
+
+func (r fakeEvalRunner) Run([]string) error {
+	return r.err
 }
 
 type fakeExitError struct {
@@ -35,255 +31,58 @@ func (e fakeExitError) ExitCode() int {
 	return e.code
 }
 
-func TestRunEvalDelegatesOracle(t *testing.T) {
-	path := t.TempDir()
-	var stderr bytes.Buffer
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, &stderr)
-	opts.debug = true
-
-	err := runEval(opts, []string{
-		path,
-		"--extra",
-		"value",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	wantPreflight := recordedCommand{
-		name: "/bin/runme",
-		args: []string{"harbor", "stdio"},
-	}
-	wantDelegate := recordedCommand{
-		name: "/bin/runme-harbor",
-		args: []string{
-			"run",
-			mustAbs(t, path),
-			"--agent", "oracle",
-			"--jobs-dir", defaultJobsDir(t),
-			"--debug",
-			"--",
-			"--extra",
-			"value",
-		},
-	}
-	if !sameCommand(calls[0], wantPreflight) {
-		t.Fatalf("preflight = %#v, want %#v", calls[0], wantPreflight)
-	}
-	if !sameCommand(calls[1], wantDelegate) {
-		t.Fatalf("delegate = %#v, want %#v", calls[1], wantDelegate)
-	}
-	wantDebug := shellCommandString(append([]string{wantDelegate.name}, wantDelegate.args...)) + "\n"
-	if stderr.String() != wantDebug {
-		t.Fatalf("debug output = %q, want %q", stderr.String(), wantDebug)
-	}
-	if got := envValue(calls[1].env, "RUNME_BIN"); got != "/bin/runme" {
-		t.Fatalf("RUNME_BIN = %q, want /bin/runme", got)
-	}
-}
-
-func TestRunEvalDefaultsDatasetPath(t *testing.T) {
-	tmp := t.TempDir()
-	t.Chdir(tmp)
-	if err := os.MkdirAll(defaultEvalDatasetPath, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-
-	err := runEval(opts, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	want := []string{
-		"run",
-		defaultDatasetPath(t),
-		"--agent", "oracle",
-		"--jobs-dir", defaultJobsDir(t),
-	}
-	if !reflect.DeepEqual(calls[1].args, want) {
-		t.Fatalf("args = %#v, want %#v", calls[1].args, want)
-	}
-	if calls[1].workingDir != defaultEvalBaseDir(tmp) {
-		t.Fatalf("workingDir = %q, want %q", calls[1].workingDir, defaultEvalBaseDir(tmp))
-	}
-}
-
-func TestRunEvalDefaultsDatasetPathWithPassthrough(t *testing.T) {
-	tmp := t.TempDir()
-	t.Chdir(tmp)
-	if err := os.MkdirAll(defaultEvalDatasetPath, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-
-	err := runEval(opts, []string{"--model", "haiku"})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	want := []string{
-		"run",
-		defaultDatasetPath(t),
-		"--agent", "oracle",
-		"--jobs-dir", defaultJobsDir(t),
-		"--",
+func TestEvalCmdPassesOptionsToHarborRunner(t *testing.T) {
+	var gotOpts harbor.EvalOptions
+	var gotArgs []string
+	cmd := evalCmd()
+	var stdout, stderr bytes.Buffer
+	cmd.SetOut(&stdout)
+	cmd.SetErr(&stderr)
+	cmd.SetArgs([]string{
+		"dataset",
+		"--agent", "codex",
+		"--task-dir", "simple-agent",
+		"--jobs-dir", "jobs",
+		"--yes",
 		"--model", "haiku",
-	}
-	if !reflect.DeepEqual(calls[1].args, want) {
-		t.Fatalf("args = %#v, want %#v", calls[1].args, want)
-	}
-	if calls[1].workingDir != defaultEvalBaseDir(tmp) {
-		t.Fatalf("workingDir = %q, want %q", calls[1].workingDir, defaultEvalBaseDir(tmp))
-	}
-}
-
-func TestRunEvalDefaultsUseGitRoot(t *testing.T) {
-	repoRoot := t.TempDir()
-	if _, err := git.PlainInit(repoRoot, false); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.MkdirAll(filepath.Join(repoRoot, defaultEvalDatasetPath), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	nested := filepath.Join(repoRoot, "nested", "dir")
-	if err := os.MkdirAll(nested, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	t.Chdir(nested)
-	baseDir := defaultEvalBaseDir(nested)
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-
-	err := runEval(opts, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	want := []string{
-		"run",
-		filepath.Join(baseDir, defaultEvalDatasetPath),
-		"--agent", "oracle",
-		"--jobs-dir", filepath.Join(baseDir, defaultEvalJobsDir),
-	}
-	if !reflect.DeepEqual(calls[1].args, want) {
-		t.Fatalf("args = %#v, want %#v", calls[1].args, want)
-	}
-	if calls[1].workingDir != baseDir {
-		t.Fatalf("workingDir = %q, want %q", calls[1].workingDir, baseDir)
-	}
-}
-
-func TestRunEvalExplicitDatasetUsesCwdAndDefaultJobsUseGitRoot(t *testing.T) {
-	repoRoot := t.TempDir()
-	if _, err := git.PlainInit(repoRoot, false); err != nil {
-		t.Fatal(err)
-	}
-	nested := filepath.Join(repoRoot, "nested", "dir")
-	dataset := filepath.Join(nested, "custom-dataset")
-	if err := os.MkdirAll(dataset, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	t.Chdir(nested)
-	baseDir := defaultEvalBaseDir(nested)
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-
-	err := runEval(opts, []string{"./custom-dataset"})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	want := []string{
-		"run",
-		dataset,
-		"--agent", "oracle",
-		"--jobs-dir", filepath.Join(baseDir, defaultEvalJobsDir),
-	}
-	if !reflect.DeepEqual(calls[1].args, want) {
-		t.Fatalf("args = %#v, want %#v", calls[1].args, want)
-	}
-	if calls[1].workingDir != baseDir {
-		t.Fatalf("workingDir = %q, want %q", calls[1].workingDir, baseDir)
-	}
-}
-
-func TestRunEvalExplicitJobsDirIsUnchanged(t *testing.T) {
-	repoRoot := t.TempDir()
-	if _, err := git.PlainInit(repoRoot, false); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.MkdirAll(filepath.Join(repoRoot, defaultEvalDatasetPath), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	nested := filepath.Join(repoRoot, "nested", "dir")
-	if err := os.MkdirAll(nested, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	t.Chdir(nested)
-	baseDir := defaultEvalBaseDir(nested)
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-	opts.jobsDir = "custom/jobs"
-	opts.jobsDirExplicit = true
-
-	err := runEval(opts, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	want := []string{
-		"run",
-		filepath.Join(baseDir, defaultEvalDatasetPath),
-		"--agent", "oracle",
-		"--jobs-dir", "custom/jobs",
-	}
-	if !reflect.DeepEqual(calls[1].args, want) {
-		t.Fatalf("args = %#v, want %#v", calls[1].args, want)
-	}
-	if calls[1].workingDir != baseDir {
-		t.Fatalf("workingDir = %q, want %q", calls[1].workingDir, baseDir)
-	}
-}
-
-func TestRunEvalDelegatesCodexAndClaudeOptions(t *testing.T) {
-	for _, tt := range []struct {
-		name  string
-		agent string
-	}{
-		{name: "codex", agent: "codex"},
-		{name: "claude-code", agent: "claude-code"},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			path := t.TempDir()
-			var calls []recordedCommand
-			opts := testEvalOptions(t, &calls, io.Discard)
-			opts.agent = tt.agent
-			opts.taskDir = "simple-agent"
-			opts.jobsDir = "jobs"
-			opts.jobsDirExplicit = true
-			opts.yes = true
-
-			err := runEval(opts, []string{path})
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			want := []string{
-				"run",
-				mustAbs(t, path),
-				"--agent", tt.agent,
-				"--jobs-dir", "jobs",
-				"--task-dir", "simple-agent",
-				"-y",
-			}
-			if !reflect.DeepEqual(calls[1].args, want) {
-				t.Fatalf("args = %#v, want %#v", calls[1].args, want)
-			}
+		"--env", "docker",
+		"--runme-bin", "/bin/runme",
+		"--runme-arg", "--skip-prompts",
+		"--runme-harbor-bin", "/bin/runme-harbor",
+		"--debug",
+		"--",
+		"--extra",
+	})
+	restoreEvalRunner(t, func(opts harbor.EvalOptions) evalRunner {
+		gotOpts = opts
+		return evalRunnerFunc(func(args []string) error {
+			gotArgs = append([]string(nil), args...)
+			return nil
 		})
+	})
+
+	if err := cmd.Execute(); err != nil {
+		t.Fatal(err)
+	}
+
+	if !reflect.DeepEqual(gotArgs, []string{"dataset", "--extra"}) {
+		t.Fatalf("args = %#v", gotArgs)
+	}
+	if gotOpts.Agent != "codex" ||
+		gotOpts.TaskDir != "simple-agent" ||
+		gotOpts.JobsDir != "jobs" ||
+		!gotOpts.JobsDirExplicit ||
+		!gotOpts.Yes ||
+		gotOpts.Model != "haiku" ||
+		gotOpts.Env != "docker" ||
+		gotOpts.RunmeBin != "/bin/runme" ||
+		!reflect.DeepEqual(gotOpts.RunmeArgs, []string{"--skip-prompts"}) ||
+		gotOpts.RunmeHarborBin != "/bin/runme-harbor" ||
+		!gotOpts.Debug ||
+		gotOpts.Stdout != &stdout ||
+		gotOpts.Stderr != &stderr ||
+		!gotOpts.Preflight {
+		t.Fatalf("options = %#v", gotOpts)
 	}
 }
 
@@ -339,307 +138,6 @@ func TestEvalCmdHelpIncludesDefaultDatasetPath(t *testing.T) {
 	}
 }
 
-func TestRunEvalDelegatesModel(t *testing.T) {
-	path := t.TempDir()
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-	opts.model = "haiku"
-
-	err := runEval(opts, []string{path})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	want := []string{
-		"run",
-		mustAbs(t, path),
-		"--agent", "oracle",
-		"--jobs-dir", defaultJobsDir(t),
-		"--",
-		"--model", "haiku",
-	}
-	if !reflect.DeepEqual(calls[1].args, want) {
-		t.Fatalf("args = %#v, want %#v", calls[1].args, want)
-	}
-}
-
-func TestRunEvalPreservesPassthroughModel(t *testing.T) {
-	path := t.TempDir()
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-
-	err := runEval(opts, []string{path, "--model", "haiku"})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	want := []string{
-		"run",
-		mustAbs(t, path),
-		"--agent", "oracle",
-		"--jobs-dir", defaultJobsDir(t),
-		"--",
-		"--model", "haiku",
-	}
-	if !reflect.DeepEqual(calls[1].args, want) {
-		t.Fatalf("args = %#v, want %#v", calls[1].args, want)
-	}
-}
-
-func TestRunEvalRejectsDuplicateModel(t *testing.T) {
-	for _, tt := range []struct {
-		name        string
-		passthrough []string
-	}{
-		{name: "separate arg", passthrough: []string{"--model", "sonnet"}},
-		{name: "equals arg", passthrough: []string{"--model=sonnet"}},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			path := t.TempDir()
-			var calls []recordedCommand
-			opts := testEvalOptions(t, &calls, io.Discard)
-			opts.model = "haiku"
-
-			err := runEval(opts, append([]string{path}, tt.passthrough...))
-			if err == nil {
-				t.Fatal("expected error")
-			}
-			if !strings.Contains(err.Error(), "--model cannot be used together with passthrough --model") {
-				t.Fatalf("error = %q", err.Error())
-			}
-			if len(calls) != 0 {
-				t.Fatalf("calls = %#v, want none", calls)
-			}
-		})
-	}
-}
-
-func TestRunEvalDelegatesRunmeEnvAlias(t *testing.T) {
-	path := t.TempDir()
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-	opts.env = "runme"
-
-	err := runEval(opts, []string{path})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	wantPreflight := recordedCommand{
-		name: "/bin/runme",
-		args: []string{"harbor", "stdio"},
-	}
-	wantDelegate := []string{
-		"run",
-		mustAbs(t, path),
-		"--agent", "oracle",
-		"--jobs-dir", defaultJobsDir(t),
-		"--env", "runme",
-	}
-	if !sameCommand(calls[0], wantPreflight) {
-		t.Fatalf("preflight = %#v, want %#v", calls[0], wantPreflight)
-	}
-	if !reflect.DeepEqual(calls[1].args, wantDelegate) {
-		t.Fatalf("args = %#v, want %#v", calls[1].args, wantDelegate)
-	}
-}
-
-func TestRunEvalDelegatesNonRunmeEnvWithoutPreflight(t *testing.T) {
-	path := t.TempDir()
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-	opts.env = "docker"
-	opts.agent = "codex"
-
-	err := runEval(opts, []string{path})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	want := []string{
-		"run",
-		mustAbs(t, path),
-		"--agent", "codex",
-		"--jobs-dir", defaultJobsDir(t),
-		"--env", "docker",
-	}
-	if len(calls) != 1 {
-		t.Fatalf("calls = %#v, want delegate only", calls)
-	}
-	if !reflect.DeepEqual(calls[0].args, want) {
-		t.Fatalf("args = %#v, want %#v", calls[0].args, want)
-	}
-}
-
-func TestRunEvalDoesNotStageNonDockerWorkdir(t *testing.T) {
-	workspace := t.TempDir()
-	t.Chdir(workspace)
-	dataset, workdir, target := makeHarborDockerDataset(t, workspace, "/app/source/workdir")
-	if err := os.WriteFile(filepath.Join(workdir, "keep.txt"), []byte("keep"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-	opts.env = "podman"
-
-	err := runEval(opts, []string{dataset})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if _, err := os.Stat(target); !os.IsNotExist(err) {
-		t.Fatalf("target stat err = %v, want not exist", err)
-	}
-}
-
-func TestRunEvalRejectsPassthroughEnvironmentFlags(t *testing.T) {
-	for _, tt := range []struct {
-		name        string
-		passthrough []string
-	}{
-		{name: "env separate", passthrough: []string{"--env", "docker"}},
-		{name: "env equals", passthrough: []string{"--env=docker"}},
-		{name: "env shorthand separate", passthrough: []string{"-e", "docker"}},
-		{name: "env shorthand joined", passthrough: []string{"-edocker"}},
-		{name: "import path separate", passthrough: []string{"--environment-import-path", "pkg:Env"}},
-		{name: "import path equals", passthrough: []string{"--environment-import-path=pkg:Env"}},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			path := t.TempDir()
-			var calls []recordedCommand
-			opts := testEvalOptions(t, &calls, io.Discard)
-
-			err := runEval(opts, append([]string{path}, tt.passthrough...))
-			if err == nil {
-				t.Fatal("expected error")
-			}
-			if !strings.Contains(err.Error(), "use runme eval --env") {
-				t.Fatalf("error = %q", err.Error())
-			}
-			if len(calls) != 0 {
-				t.Fatalf("calls = %#v, want none", calls)
-			}
-		})
-	}
-}
-
-func TestRunEvalUsesEnvAndRunmeArgs(t *testing.T) {
-	envHarbor := makeExecutable(t, "env-runme-harbor")
-	flagHarbor := makeExecutable(t, "flag-runme-harbor")
-	t.Setenv("RUNME_HARBOR_BIN", envHarbor)
-	t.Setenv("RUNME_ARGS", "--existing")
-	path := t.TempDir()
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-	opts.runmeHarbor = flagHarbor
-	opts.runmeBin = "/flag/runme"
-	opts.runmeArgs = []string{"--chdir", "/tmp/with space", "--label=can't-$expand`this`\\path"}
-
-	err := runEval(opts, []string{path})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if calls[1].name != envHarbor {
-		t.Fatalf("runme-harbor = %q, want env override", calls[1].name)
-	}
-	if got := envValue(calls[1].env, "RUNME_BIN"); got != "/flag/runme" {
-		t.Fatalf("RUNME_BIN = %q, want /flag/runme", got)
-	}
-	if got := envValue(calls[1].env, "RUNME_ARGS"); got != "--chdir '/tmp/with space' '--label=can'\\''t-$expand`this`\\path'" {
-		t.Fatalf("RUNME_ARGS = %q", got)
-	}
-}
-
-func TestRunEvalPreservesExistingRunmeArgs(t *testing.T) {
-	t.Setenv("RUNME_ARGS", "--existing")
-	path := t.TempDir()
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-
-	err := runEval(opts, []string{path})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if got := envValue(calls[1].env, "RUNME_ARGS"); got != "--existing" {
-		t.Fatalf("RUNME_ARGS = %q, want preserved env", got)
-	}
-}
-
-func TestRunEvalMissingRunmeHarbor(t *testing.T) {
-	path := t.TempDir()
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-	opts.lookPath = func(string) (string, error) {
-		return "", os.ErrNotExist
-	}
-
-	err := runEval(opts, []string{path})
-	if err == nil {
-		t.Fatal("expected error")
-	}
-	if !strings.Contains(err.Error(), "uv tool install runme-harbor") {
-		t.Fatalf("error = %q", err.Error())
-	}
-	if len(calls) != 0 {
-		t.Fatalf("calls = %#v, want none", calls)
-	}
-}
-
-func TestRunEvalDelegatesUnknownAgent(t *testing.T) {
-	path := t.TempDir()
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-	opts.agent = "bad"
-
-	err := runEval(opts, []string{path})
-	if err != nil {
-		t.Fatal(err)
-	}
-	want := []string{
-		"run",
-		mustAbs(t, path),
-		"--agent", "bad",
-		"--jobs-dir", defaultJobsDir(t),
-	}
-	if !reflect.DeepEqual(calls[1].args, want) {
-		t.Fatalf("args = %#v, want %#v", calls[1].args, want)
-	}
-}
-
-func TestRunEvalMissingPath(t *testing.T) {
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-
-	err := runEval(opts, []string{filepath.Join(t.TempDir(), "missing")})
-	if err == nil || !strings.Contains(err.Error(), "dataset path does not exist") {
-		t.Fatalf("error = %v", err)
-	}
-}
-
-func TestRunEvalMissingDefaultPath(t *testing.T) {
-	tmp := t.TempDir()
-	t.Chdir(tmp)
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-
-	err := runEval(opts, nil)
-	if err == nil {
-		t.Fatal("expected error")
-	}
-	if !strings.Contains(err.Error(), "dataset path does not exist: evals/tasks") {
-		t.Fatalf("error = %q", err.Error())
-	}
-	if !strings.Contains(err.Error(), "pass a dataset path explicitly") {
-		t.Fatalf("error = %q", err.Error())
-	}
-	if len(calls) != 0 {
-		t.Fatalf("calls = %#v, want none", calls)
-	}
-}
-
 func TestEvalCmdRejectsMultipleDatasetPaths(t *testing.T) {
 	cmd := evalCmd()
 	cmd.SetArgs([]string{"first", "second"})
@@ -656,18 +154,11 @@ func TestEvalCmdRejectsMultipleDatasetPaths(t *testing.T) {
 }
 
 func TestRunEvalPropagatesDelegateExitCode(t *testing.T) {
-	path := t.TempDir()
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-	opts.commandRun = func(name string, args []string, workingDir string, env []string, stdin io.Reader, stdout, stderr io.Writer) error {
-		calls = append(calls, recordedCommand{name: name, args: append([]string(nil), args...), workingDir: workingDir, env: env})
-		if len(calls) == 2 {
-			return fakeExitError{code: 42}
-		}
-		return nil
-	}
+	restoreEvalRunner(t, func(harbor.EvalOptions) evalRunner {
+		return fakeEvalRunner{err: fakeExitError{code: 42}}
+	})
 
-	err := runEval(opts, []string{path})
+	err := runEval(evalOptions{}, []string{"dataset"})
 	var exitErr ExitCodeError
 	if !errors.As(err, &exitErr) {
 		t.Fatalf("error = %T %v, want ExitCodeError", err, err)
@@ -677,228 +168,31 @@ func TestRunEvalPropagatesDelegateExitCode(t *testing.T) {
 	}
 }
 
-func TestRunEvalPrintsExceptionDetailsAfterHarborOutput(t *testing.T) {
-	path := t.TempDir()
-	jobsDir := filepath.Join(t.TempDir(), "jobs")
-	jobDir := filepath.Join(jobsDir, "2026-06-18__10-51-14")
-	resultPath := filepath.Join(jobDir, "result.json")
-	var stdout bytes.Buffer
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-	opts.jobsDir = jobsDir
-	opts.jobsDirExplicit = true
-	opts.stdout = &stdout
-	opts.commandRun = func(name string, args []string, workingDir string, env []string, stdin io.Reader, stdoutWriter, stderr io.Writer) error {
-		calls = append(calls, recordedCommand{name: name, args: append([]string(nil), args...), workingDir: workingDir, env: env})
-		if len(calls) == 2 {
-			_, _ = fmt.Fprintf(stdoutWriter, "Exception           Count\nHarborProtocolError 1\nResults written to %s\n", resultPath)
-			writeEvalException(t, jobDir, "end-to-end__abc/exception.txt", "HarborProtocolError: runme failed with useful context\n")
-			return fakeExitError{code: 7}
-		}
-		return nil
-	}
+func TestRunEvalMissingRunmeHarborMessage(t *testing.T) {
+	restoreEvalRunner(t, func(harbor.EvalOptions) evalRunner {
+		return fakeEvalRunner{err: harbor.ErrRunmeHarborMissing}
+	})
 
-	err := runEval(opts, []string{path})
-	var exitErr ExitCodeError
-	if !errors.As(err, &exitErr) {
-		t.Fatalf("error = %T %v, want ExitCodeError", err, err)
+	err := runEval(evalOptions{}, []string{"dataset"})
+	if err == nil {
+		t.Fatal("expected error")
 	}
-
-	output := string(ansi.Strip(stdout.Bytes()))
-	if !strings.Contains(output, "Exception           Count\nHarborProtocolError 1\nResults written to "+resultPath+"\n\nHarbor Exception Details\n") {
-		t.Fatalf("output = %q", output)
-	}
-	if !strings.Contains(output, "File: end-to-end__abc/exception.txt") {
-		t.Fatalf("output = %q", output)
-	}
-	if !strings.Contains(output, "HarborProtocolError: runme failed with useful context") {
-		t.Fatalf("output = %q", output)
+	if !strings.Contains(err.Error(), "uv tool install runme-harbor") {
+		t.Fatalf("error = %q", err.Error())
 	}
 }
 
-func TestRunEvalPrintsExceptionDetailsForDockerEnvironment(t *testing.T) {
-	workspace := t.TempDir()
-	t.Chdir(workspace)
-	dataset, _, _ := makeHarborDockerDataset(t, workspace, "/app/source/workdir")
-	jobsDir := filepath.Join(t.TempDir(), "jobs")
-	jobDir := filepath.Join(jobsDir, "2026-06-18__10-51-14")
-	resultPath := filepath.Join(jobDir, "result.json")
-	var stdout bytes.Buffer
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-	opts.env = "docker"
-	opts.jobsDir = jobsDir
-	opts.jobsDirExplicit = true
-	opts.stdout = &stdout
-	opts.commandRun = func(name string, args []string, workingDir string, env []string, stdin io.Reader, stdoutWriter, stderr io.Writer) error {
-		calls = append(calls, recordedCommand{name: name, args: append([]string(nil), args...), workingDir: workingDir, env: env})
-		_, _ = fmt.Fprintf(stdoutWriter, "Exception           Count\nHarborProtocolError 1\nResults written to %s\n", resultPath)
-		writeEvalException(t, jobDir, "end-to-end__abc/exception.txt", "ValueError: docker detail should be visible\n")
-		return fakeExitError{code: 7}
-	}
+type evalRunnerFunc func([]string) error
 
-	err := runEval(opts, []string{dataset})
-	var exitErr ExitCodeError
-	if !errors.As(err, &exitErr) {
-		t.Fatalf("error = %T %v, want ExitCodeError", err, err)
-	}
-	output := string(ansi.Strip(stdout.Bytes()))
-	if !strings.Contains(output, "Harbor Exception Details") || !strings.Contains(output, "ValueError: docker detail should be visible") {
-		t.Fatalf("output = %q", stdout.String())
-	}
+func (f evalRunnerFunc) Run(args []string) error {
+	return f(args)
 }
 
-func TestRunEvalOnlyPrintsExceptionDetailsForReportedJob(t *testing.T) {
-	path := t.TempDir()
-	jobsDir := filepath.Join(t.TempDir(), "jobs")
-	jobDir := filepath.Join(jobsDir, "current")
-	resultPath := filepath.Join(jobDir, "result.json")
-	writeEvalException(t, jobsDir, "other/attempt/exception.txt", "other detail\n")
-	writeEvalException(t, jobDir, "attempt/exception.txt", "current detail\n")
-	var stdout bytes.Buffer
-	var calls []recordedCommand
-	opts := testEvalOptions(t, &calls, io.Discard)
-	opts.jobsDir = jobsDir
-	opts.jobsDirExplicit = true
-	opts.stdout = &stdout
-	opts.commandRun = func(name string, args []string, workingDir string, env []string, stdin io.Reader, stdoutWriter, stderr io.Writer) error {
-		calls = append(calls, recordedCommand{name: name, args: append([]string(nil), args...), workingDir: workingDir, env: env})
-		if len(calls) == 2 {
-			_, _ = fmt.Fprintf(stdoutWriter, "Exception           Count\nHarborProtocolError 1\nResults written to %s\n", resultPath)
-			return fakeExitError{code: 7}
-		}
-		return nil
-	}
-
-	err := runEval(opts, []string{path})
-	var exitErr ExitCodeError
-	if !errors.As(err, &exitErr) {
-		t.Fatalf("error = %T %v, want ExitCodeError", err, err)
-	}
-	output := string(ansi.Strip(stdout.Bytes()))
-	if !strings.Contains(output, "Harbor Exception Details") || !strings.Contains(output, "current detail") {
-		t.Fatalf("output = %q", stdout.String())
-	}
-	if strings.Contains(output, "other detail") {
-		t.Fatalf("output = %q", stdout.String())
-	}
-}
-
-func testEvalOptions(t *testing.T, calls *[]recordedCommand, stderr io.Writer) evalOptions {
+func restoreEvalRunner(t *testing.T, fn func(harbor.EvalOptions) evalRunner) {
 	t.Helper()
-	return evalOptions{
-		agent:       "oracle",
-		jobsDir:     defaultEvalJobsDir,
-		runmeBin:    "/bin/runme",
-		commandRun:  recordCommand(calls),
-		lookPath:    fakeLookPath,
-		executable:  func() (string, error) { return "/bin/current-runme", nil },
-		stdout:      io.Discard,
-		stderr:      stderr,
-		preflight:   true,
-		runmeHarbor: "",
-	}
-}
-
-func recordCommand(calls *[]recordedCommand) commandRunFunc {
-	return func(name string, args []string, workingDir string, env []string, stdin io.Reader, stdout, stderr io.Writer) error {
-		*calls = append(*calls, recordedCommand{
-			name:       name,
-			args:       append([]string(nil), args...),
-			workingDir: workingDir,
-			env:        append([]string(nil), env...),
-		})
-		return nil
-	}
-}
-
-func fakeLookPath(name string) (string, error) {
-	switch name {
-	case "runme-harbor":
-		return "/bin/runme-harbor", nil
-	case "/env/runme-harbor", "/flag/runme-harbor":
-		return name, nil
-	default:
-		return "", os.ErrNotExist
-	}
-}
-
-func sameCommand(got, want recordedCommand) bool {
-	return got.name == want.name && reflect.DeepEqual(got.args, want.args)
-}
-
-func mustAbs(t *testing.T, path string) string {
-	t.Helper()
-	abs, err := filepath.Abs(path)
-	if err != nil {
-		t.Fatal(err)
-	}
-	return abs
-}
-
-func defaultJobsDir(t *testing.T) string {
-	t.Helper()
-	cwd, err := os.Getwd()
-	if err != nil {
-		t.Fatal(err)
-	}
-	return filepath.Join(defaultEvalBaseDir(cwd), defaultEvalJobsDir)
-}
-
-func defaultDatasetPath(t *testing.T) string {
-	t.Helper()
-	cwd, err := os.Getwd()
-	if err != nil {
-		t.Fatal(err)
-	}
-	return filepath.Join(defaultEvalBaseDir(cwd), defaultEvalDatasetPath)
-}
-
-func envValue(env []string, key string) string {
-	prefix := key + "="
-	for _, item := range env {
-		if strings.HasPrefix(item, prefix) {
-			return strings.TrimPrefix(item, prefix)
-		}
-	}
-	return ""
-}
-
-func makeExecutable(t *testing.T, name string) string {
-	t.Helper()
-	path := filepath.Join(t.TempDir(), name)
-	if err := os.WriteFile(path, []byte("#!/bin/sh\n"), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	return path
-}
-
-func makeHarborDockerDataset(t *testing.T, workspace string, remoteWorkdir string) (string, string, string) {
-	t.Helper()
-	dataset := filepath.Join(workspace, "evals", "tasks")
-	task := filepath.Join(dataset, "example-task")
-	workdir := filepath.Join(workspace, "source", "workdir")
-	target := filepath.Join(task, "environment", "workdir")
-	if err := os.MkdirAll(workdir, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.MkdirAll(filepath.Join(task, "environment"), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	config := "schema_version = \"1.1\"\n\n[environment]\nworkdir = \"" + remoteWorkdir + "\"\n"
-	if err := os.WriteFile(filepath.Join(task, "task.toml"), []byte(config), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	return dataset, workdir, target
-}
-
-func writeEvalException(t *testing.T, jobsDir, relativePath, content string) {
-	t.Helper()
-	path := filepath.Join(jobsDir, relativePath)
-	if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(path, []byte(content), 0o644); err != nil {
-		t.Fatal(err)
-	}
+	previous := newEvalRunner
+	newEvalRunner = fn
+	t.Cleanup(func() {
+		newEvalRunner = previous
+	})
 }
